requirements.txt
Flask==2.2.3
joblib==1.2.0
keras==2.12.0
nltk==3.8.1
numpy==1.23.5
Pillow==9.5.0
scikit_learn==1.2.2
scipy==1.10.1
spacy==3.5.1
tensorflow==2.12.0
transformers==4.27.4
